Add files via upload
Code which is almost complete, predicting the future of football players whether the player should stay on the club or look for better option by analysing the performance of the players from previous season. Working on providing more data from the previous seasons. This dataset only has the players from premier league season 2023/24.